inventory-sheet/requirements.txt
xlsxwriter
PyYAML
boto3